fix: tiny change

+++ src/code/conv/conv_3ntt.hpp
@@ -27,7 +27,7 @@ vec<u64> conv_3ntt(vec<u64> const &l, vec<u64> const &r, u64 mod, u32 ans_size) 
   return ret;
 }
 template <class mint0, class mint1, class mint2>
-vec<u64> conv_3ntt(vec<u64> const &l, vec<u64> const &r, u64 mod) { return conv_3ntt<mint0, mint1, mint2>(l, r, mod, l.size() + r.size() - 1); }
+vec<u64> conv_3ntt(vec<u64> const &l, vec<u64> const &r, u64 mod) { return conv_3ntt<mint0, mint1, mint2>(l, r, mod, u32(l.size() + r.size() - 1)); }
 template <class mint, class mint0, class mint1, class mint2>
 vec<mint> conv_3ntt(vec<mint> const &l, vec<mint> const &r, u32 ans_size) {
   vec<u64> l_(l.size()), r_(r.size());
@@ -39,7 +39,7 @@ vec<mint> conv_3ntt(vec<mint> const &l, vec<mint> const &r, u32 ans_size) {
   return res;
 }
 template <class mint, class mint0, class mint1, class mint2>
-vec<mint> conv_3ntt(vec<mint> const &l, vec<mint> const &r) { return conv_3ntt<mint, mint0, mint1, mint2>(l, r, l.size() + r.size() - 1); }
+vec<mint> conv_3ntt(vec<mint> const &l, vec<mint> const &r) { return conv_3ntt<mint, mint0, mint1, mint2>(l, r, u32(l.size() + r.size() - 1)); }
 
 }  // namespace tifa_libs::math
