+++ src/code/conv/conv_ntt32.hpp
@@ -9,7 +9,7 @@ namespace tifa_libs::math {
 template <class mint>
 vec<mint> conv_ntt32(vec<mint> l, vec<mint> r, u32 ans_size) {
   static NTT32<mint> ntt;
-  ntt.bzr(std::max({l.size(), r.size(), std::min(l.size() + r.size() - 1, ans_size)}));
+  ntt.bzr(std::max({(u32)l.size(), (u32)r.size(), std::min(u32(l.size() + r.size() - 1), ans_size)}));
   ntt.dif(l);
   ntt.dif(r);
   for (u32 i = 0; i < ntt.size(); ++i) l[i] *= r[i];
@@ -18,7 +18,7 @@ vec<mint> conv_ntt32(vec<mint> l, vec<mint> r, u32 ans_size) {
   return l;
 }
 template <class mint>
-vec<mint> conv_ntt32(vec<mint> const &l, vec<mint> const &r) { return conv_ntt32(l, r, l.size() + r.size() - 1); }
+vec<mint> conv_ntt32(vec<mint> const &l, vec<mint> const &r) { return conv_ntt32(l, r, u32(l.size() + r.size() - 1)); }
 template <class mint, class T = u32>
 vec<mint> conv_ntt32(vec<T> const &l, vec<T> const &r, u32 ans_size) {
   static_assert(sizeof(T) <= 32);
@@ -30,7 +30,7 @@ vec<mint> conv_ntt32(vec<T> const &l, vec<T> const &r, u32 ans_size) {
   return conv_ntt32(l_, r_, ans_size);
 }
 template <class mint, class T = u32>
-vec<mint> conv_ntt32(vec<T> const &l, vec<T> const &r) { return conv_ntt32<mint, T>(l, r, l.size() + r.size() - 1); }
+vec<mint> conv_ntt32(vec<T> const &l, vec<T> const &r) { return conv_ntt32<mint, T>(l, r, u32(l.size() + r.size() - 1)); }
 
 }  // namespace tifa_libs::math
