+++ src/code/poly/polydata_d3ntt.hpp
@@ -19,7 +19,7 @@ struct polydata_d3ntt {
   explicit constexpr polydata_d3ntt(vec<mint> const &v) : d(v) {}
 
   void conv(polydata_d3ntt const &r, u32 ans_size) { d = ans_size < 32 ? conv_naive(d, r.d, ans_size) : conv_3ntt<mint, mint0, mint1, mint2>(d, r.d, ans_size); }
-  void conv(polydata_d3ntt const &r) { conv(r, d.size() + r.d.size() - 1); }
+  void conv(polydata_d3ntt const &r) { conv(r, u32(d.size() + r.d.size() - 1)); }
 };
 
 }  // namespace tifa_libs::math
